refactor out components for GameList
