+++ src/routes/index/components/inProgressGame/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './InProgressGame';
