+++ src/routes/index/components/openGame/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './OpenGame';
